Se ha implementado el protocolo de autorización OAuth2 y el proveedor de google, modificados los privilegios e implementado validaciones faltantes.

+++ src/main/java/com/example/demo/DTOs/DireccionDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class DireccionDTO {
 
-    private Integer idDireccion;
+    private Long idDireccion;
     private String tipoVia;
     private String numeroViaPrincipal;
     private String letraViaPrincipal;
